21-09-24

+++ algo/src/BOJ/Bronze/BOJ2748.java
@@ -5,8 +5,8 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 
 public class BOJ2748 {
-    static int arr[];
-    static int fibo(int n){
+    static long arr[];
+    static long fibo(int n){
         if(n>=2&&arr[n]==0){
             arr[n]=fibo(n-2)+fibo(n-1);
         }
@@ -15,9 +15,10 @@ public class BOJ2748 {
     public static void main(String[] args) throws IOException {
         BufferedReader br =new BufferedReader(new InputStreamReader(System.in));
         int T =Integer.parseInt(br.readLine());
-        arr= new int[91];
+        arr= new long[91];
         arr[0]=0;
         arr[1]=1;
+
 
 
         System.out.println(fibo(T));
